Move NewMoney and NewDateTime to proto.go file
